refactor: Extract `buildRichText` function into separate utils file
